An arm created with two servos controlled by 4 buttons via arduino.
An intro to learning how to program an arduino while also learning how
circuits work. The arm made is up of 3 cardboard pieces (a base, and 2
rectangular pieces) with two servos that serve as the moving joints.
The arm can go up/down and left/right controlled through 4 button
circuited on a breadboard.